refactor: decouple CLI UI, fix PDFium thread safety, smart favicon package output, simplify batch pipeline with Rayon, and resolve Clippy unnecessary cast

+++ src/converter/pdf_conv.rs
@@ -43,7 +43,7 @@ impl PdfConverter {
                 reason: format!("Failed to parse PDF document: {:?}", e),
             })?;
 
-        let total_pages = document.pages().len() as u16;
+        let total_pages = document.pages().len();
         if total_pages == 0 {
             return Err(MartiniError::InvalidInputData {
                 reason: "PDF document has no pages".to_string(),
